requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
openai==1.3.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
openpyxl==3.1.2
pandas==2.1.3 